change new filter from CheckboxSelectMultiple to ChoiceField, move code from leaderboad views to mixin in mixins.py
