+++ contributors/forms/__init__.py
@@ -1,2 +1,1 @@
 from contributors.forms.admin_forms import OrgNamesForm, RepoNamesForm
-from contributors.forms.forms import TableSortSearchForm
