+++ contributors/views/leaderboard_issues.py
@@ -1,11 +1,17 @@
 from django.views import generic
 
-from contributors.forms.forms import LeaderboardCombinedSearchForm
 from contributors.models import Contributor
-from contributors.views.mixins import TableSortSearchAndPaginationMixin
+from contributors.views.mixins import (
+    LeaderboardQueryMixin,
+    TableSortSearchAndPaginationMixin,
+)
 
 
-class ListView(TableSortSearchAndPaginationMixin, generic.ListView):
+class ListView(
+    LeaderboardQueryMixin,
+    TableSortSearchAndPaginationMixin,
+    generic.ListView,
+):
     """List of leaders among contributors by issues."""
 
     queryset = Contributor.objects.visible().with_contributions()
@@ -18,26 +24,3 @@ class ListView(TableSortSearchAndPaginationMixin, generic.ListView):
     searchable_fields = ('login', 'name')
     ordering = sortable_fields[0]
     paginate_by = 100
-
-    def get_context_data(self, **kwargs):
-        """Get search form by organizations."""
-        context = super().get_context_data(**kwargs)
-        context['form_org'] = LeaderboardCombinedSearchForm(self.request.GET)
-        return context
-
-    def get_queryset(self):  # noqa: WPS615
-        """Get filter queryset."""
-        queryset = super().get_queryset().prefetch_related(
-            'contributors__organization',
-        )
-        form = LeaderboardCombinedSearchForm(self.request.GET)
-        if form.is_valid():
-            organizations = form.cleaned_data['organizations']
-            if organizations:
-                queryset = queryset.filter(
-                    contributors__organization__name__exact=organizations,
-                )
-            is_hexlet_stuff = form.cleaned_data['is_hexlet_stuff']
-            if is_hexlet_stuff:
-                queryset = queryset.filter(user__is_staff=True)
-        return queryset
